# docs/swagger.yaml
openapi: 3.0.0
info:
  title: TCSS-460 Message API
  version: 1.0.0
  description: |
    Educational REST API demonstrating modern Node.js/Express/TypeScript patterns for TCSS-460.

    This API showcases:
    - RESTful message CRUD operations with PostgreSQL database
    - HTTP method semantics (GET, POST, PATCH, DELETE)
    - Request parameter validation (query, path, body)
    - Input validation and sanitization with express-validator
    - Standardized response formats with success/error handling
    - Database transaction management
    - Type-safe TypeScript implementations

    **Learning Objectives:**
    - Understand RESTful API design principles
    - Practice HTTP protocol fundamentals
    - Implement proper input validation
    - Create consistent API responses
    - Master database CRUD operations
    - Document APIs for maintainability
  contact:
    name: TCSS-460 Course
    email: cfb3@uw.edu
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:4000
    description: Development server
  - url: http://localhost:5000
    description: Alternative development port
...
paths:
...
  /doc/{filename}:
    get:
      tags:
        - Documentation
      summary: View rendered documentation file
      description: Returns markdown documentation rendered as HTML with syntax highlighting
      parameters:
        - name: filename
          in: path
          required: true
          description: Name of the markdown file to view
          schema:
            type: string
            pattern: '^[a-zA-Z0-9_-]+\.md$'
            example: http-fundamentals.md
      responses:
        '200':
          description: Rendered documentation
          content:
            text/html:
              schema:
                type: string
        '404':
          description: Documentation file not found
...